+++ aclogview/Packets.cs
@@ -87,7 +87,8 @@ public class Util {
                 str += Encoding.Unicode.GetString(binaryReader.ReadBytes(2));
             }
         }
-        readToAlign(binaryReader);
+        // Note: I had to comment this out to avoid alignment issues. (Slushnas)
+        //readToAlign(binaryReader);
         return str;
     }